worker/requirements.txt
pika==1.3.2
pdfplumber==0.11.0
psycopg2-binary==2.9.9
openai==1.12.0
httpx==0.26.0
requests==2.31.0
